tests/wpt/tests/referrer-policy/css-integration/child-css/internal-import-stylesheet-with-differentorigin-base-url.tentative.html

<!DOCTYPE html>
<title>CSS integration - Child css from internal stylesheet</title>
<link rel="help" href="https://crbug.com/1158645" />

<head>
  <meta name="referrer" content="origin">
</head>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<!-- Common global functions for referrer-policy tests. -->
<script src="/common/security-features/resources/common.sub.js"></script>

<!-- This has to follow the <script> tags, or it will make the .js files
fail to load. -->
<base href="http://other-site.example" />

<script>
  promise_test(function (css_test) {
      let id = token();
      let url_prefix = location.protocol + "//www1." + location.hostname +
        ":" + location.port;
      let css_url = url_prefix +
        "/common/security-features/subresource/stylesheet.py?id=" + id +
        "&stylesheet-only";
      let check_url = url_prefix +
        "/common/security-features/subresource/stylesheet.py" +
        "?id=" + id + "&report-headers";

      let style = document.createElement("style");
      style.type = 'text/css';
      style.appendChild(document.createTextNode("@import url('" + css_url +
        "');"));
      document.head.appendChild(style);
      return timeoutPromise(css_test, 1000)
        .then(() => requestViaXhr(check_url))
        .then(function (message) {
          assert_own_property(message, "headers");
          assert_own_property(message, "referrer");
          assert_equals(message.referrer, location.origin + "/");
        });
    },
    "A child stylesheet from inline style should use the document's URL, not " +
    "the document's base URL, as its referrer."
  );
</script>
